Move things around to make it more standard, move changelog to a different file,
