+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+Changelog:
+-----
+* 2025-05: Support way order in relataion, improve typescript typeings, added CI for vitest tests
+* 2025-03-03: fixed a few bugs around non-*Polygon features processing and convert it to a pure TypeScript project (codes published to [NPM](https://npmjs.org/) have become TSC compiled from now on).
+* 2019-2025: quite a few contributors had fixed quite a few bugs and also contributed some good ideas - thanks for their efforts!
+* Late 2018: this repository was created and the initial version was pushed to github - almost around specific usages for myself.
