+++ src/osm-object.ts
