+++ src/way.ts
