+++ tsconfig.json
@@ -6,7 +6,7 @@
         "esModuleInterop": true,
         "allowSyntheticDefaultImports": true,
         "strictNullChecks": true,
-        "outDir": "../dist/node",
+        "outDir": "dist/node",
         "strict": true,
         "target": "es2017"
     },
